+++ test/config/default-array-merge.js
@@ -1,0 +1,12 @@
+
+var defer = require('../../defer').deferConfig;
+
+var config = {
+  arrayMerging: {
+    emptyArray : [],
+    oneItem   : ['one'],
+    removeMe  : ['some value'],
+  },
+};
+
+module.exports = config;