+++ test/config/local-array-merge.js
@@ -1,0 +1,10 @@
+
+var config = {
+  arrayMerging: {
+    emptyArray : ['not empty anymore'], // override empty array with a value
+    oneItem    : ['replaced'], // Replace one item with another.
+    removeMe   : [], // Replace a full array with an empty one 
+  },
+};
+
+module.exports = config;