+++ History.md
@@ -6,6 +6,7 @@
     to the main configuration object, so it doesn't have to be passed into the function. (@markstos)
   * new defer sub-module introduced in 1.8.0 can now be accessed by require('config/defer')
     For usage, see: https://github.com/lorenwest/node-config/wiki/Configuration-Files#javascript-module---js
+  * Add test coverage for array merging cases. (@markstos)
 
 1.8.0 / 2014-11-13
 ==================